chore: fixed working-directory in tests
