+++ atlasaction/bitbucket_test.go
@@ -5,8 +5,8 @@ import (
 	"path/filepath"
 	"testing"
 
-	"ariga.io/atlas-go-sdk/atlasexec"
 	"github.com/rogpeppe/go-internal/testscript"
+	"github.com/stretchr/testify/require"
 )
 
 func TestBitbucketPipe(t *testing.T) {
@@ -14,6 +14,8 @@ func TestBitbucketPipe(t *testing.T) {
 		actions = "actions"
 		outputs = filepath.Join("outputs.sh")
 	)
+	wd, err := os.Getwd()
+	require.NoError(t, err)
 	testscript.Run(t, testscript.Params{
 		Dir: filepath.Join("testdata", "bitbucket"),
 		Setup: func(e *testscript.Env) (err error) {
@@ -21,20 +23,13 @@ func TestBitbucketPipe(t *testing.T) {
 			if err := os.Mkdir(dir, 0700); err != nil {
 				return err
 			}
+			e.Setenv("MOCK_ATLAS", filepath.Join(wd, "mock-atlas.sh"))
 			e.Setenv("BITBUCKET_PIPELINE_UUID", "fbfb4205-c666-42ed-983a-d27f47f2aad2")
 			e.Setenv("BITBUCKET_PIPE_STORAGE_DIR", dir)
 			e.Setenv("BITBUCKET_CLONE_DIR", e.WorkDir)
-			c, err := atlasexec.NewClient(e.WorkDir, "atlas")
-			if err != nil {
-				return err
-			}
-			// Create a new actions for each test.
-			e.Values[atlasKey{}] = &atlasClient{c}
 			return nil
 		},
 		Cmds: map[string]func(ts *testscript.TestScript, neg bool, args []string){
-			"atlas-action": atlasAction,
-			"mock-atlas":   mockAtlasOutput,
 			"output": func(ts *testscript.TestScript, neg bool, args []string) {
 				if len(args) == 0 {
 					_, err := os.Stat(ts.MkAbs(outputs))
