+++ atlasaction/circleci_action_test.go
@@ -13,7 +13,6 @@ import (
 	"testing"
 
 	"ariga.io/atlas-action/atlasaction"
-	"ariga.io/atlas-go-sdk/atlasexec"
 	"github.com/rogpeppe/go-internal/testscript"
 	"github.com/stretchr/testify/require"
 )
@@ -63,6 +62,8 @@ func TestCircleCI(t *testing.T) {
 		actions = "actions"
 		output  = filepath.Join(actions, "output.txt")
 	)
+	wd, err := os.Getwd()
+	require.NoError(t, err)
 	testscript.Run(t, testscript.Params{
 		Dir: filepath.Join("testdata", "circleci"),
 		Setup: func(e *testscript.Env) (err error) {
@@ -70,21 +71,14 @@ func TestCircleCI(t *testing.T) {
 			if err := os.Mkdir(dir, 0700); err != nil {
 				return err
 			}
+			e.Setenv("MOCK_ATLAS", filepath.Join(wd, "mock-atlas.sh"))
 			e.Setenv("CIRCLECI", "true")
 			e.Setenv("CIRCLE_PROJECT_REPONAME", "atlas-orb")
 			e.Setenv("CIRCLE_SHA1", "1234567890")
 			e.Setenv("BASH_ENV", filepath.Join(dir, "output.txt"))
-			c, err := atlasexec.NewClient(e.WorkDir, "atlas")
-			if err != nil {
-				return err
-			}
-			// Create a new actions for each test.
-			e.Values[atlasKey{}] = &atlasClient{c}
 			return nil
 		},
 		Cmds: map[string]func(ts *testscript.TestScript, neg bool, args []string){
-			"atlas-action": atlasAction,
-			"mock-atlas":   mockAtlasOutput,
 			"output": func(ts *testscript.TestScript, neg bool, args []string) {
 				if len(args) == 0 {
 					_, err := os.Stat(ts.MkAbs(output))
