+++ atlasaction/testdata/github/migrate-lint.txtar
@@ -1,5 +1,5 @@
 # Mock the atlas command outputs
-mock-atlas $WORK/migrate-lint
+env ATLAS_PATH=$MOCK_ATLAS TEST_BATCH=./migrate-lint
 # Setup the action input variables
 env INPUT_CONFIG=file://testdata/config/atlas.hcl
 env INPUT_ENV=test
@@ -10,7 +10,7 @@ env INPUT_DIR=file://testdata/migrations
 env INPUT_DEV_URL=sqlite://file?mode=memory
 env INPUT_RUN=example
 
-atlas-action migrate/lint
+atlas-action --action=migrate/lint
 output output.txt
 
 -- migrate-lint/1/args --
