+++ atlasaction/testdata/gitlab/schema-plan.txtar
@@ -1,7 +1,7 @@
-! atlas-action schema/plan
-stderr 'the action should be run in a pull request context'
+! atlas-action --action=schema/plan
+stdout 'the action should be run in a pull request context'
 
-mock-atlas $WORK/schema-plan
+env ATLAS_PATH=$MOCK_ATLAS TEST_BATCH=./schema-plan
 env CI_MERGE_REQUEST_IID=1
 env CI_PROJECT_NAME=my-project
 env CI_PROJECT_URL=https://gitlab.com/projects/my-project
@@ -11,11 +11,16 @@ env GITLAB_USER_NAME=user
 env GITLAB_USER_ID=123
 env CI_MERGE_REQUEST_REF_PATH=$CI_PROJECT_URL/merge_requests/1
 
-atlas-action schema/plan
+atlas-action --action=schema/plan
 stdout 'Schema plan does not exist, creating a new one with name "pr-1-3RRRcLHF"'
 
 cmp comments-expected/1 comments/1
 
+cmp .env expected.env
+-- expected.env --
+link=http://test.atlasgo.cloud/schemas/141733920769/plans/210453397511
+plan=atlas://app/plans/20241010143904
+status=PENDING
 -- schema-plan/1/args --
 schema plan list --format {{ json . }} --context {"repo":"my-project","branch":"test-branch","commit":"123","url":"https://gitlab.com/projects/my-project/merge_requests/1","username":"user","userID":"123","scmType":"GITLAB"} --pending --auto-approve
